0.0.7
* Tweaks, PlayerUI update

Tweaked CameraMovementSettings according to v0.0.6 feedback. Changed PlayerUI and hotkey system. Clear is now a seperate "menu" and can be activated in the PlayerUI-MainMenu or with "X"

* Rebuilt tile placing system

This is a preperation for new features like houses. This commit is a backup as its the newest working version.

* Added houses

Houses can now be connected to every normal street at the walkway side. However, the are currently not deletd if the corresponding street is removed.

* Fallback commit

* Optimized and simplified some tile related code

This is a preperation for the update system that is for example responsible for removing houses that no longer have a valid connection.

* Implemented update system

For now, this only applies for houses.
If their street is removed, the instantly destroy themselves.

* Fixed bug

This bug occured if something was placed at the edge.